requirements.txt
azure-cognitiveservices-vision-customvision
requests
opencv-python
pillow
